mac problem solved

+++ rank3/minishell/src/utils/get_next_line/get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: juvan-to <juvan-to@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/11/14 17:16:21 by juvan-to      #+#    #+#                 */
-/*   Updated: 2023/11/03 16:19:45 by juvan-to      ########   odam.nl         */
+/*   Updated: 2023/11/05 02:15:01 by Julia         ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,8 +25,8 @@
 # include <fcntl.h>
 
 char	*get_next_line(int fd);
-char	*gnl_strjoin(char *stash, char *str, size_t index, size_t start);
-void	*calloc_gnl(size_t nelem, size_t elsize);
+char	*gnl_strjoin(char *stash, char *str, int index, int start);
+void	*calloc_gnl(int nelem, int elsize);
 int		search_newline(char *s);
 
 #endif
